refactor: 修复走查发现的一些细节问题
